requirements.txt
boto3
smtplib
ssl
google-api-python-client
